requirements.txt
# Core dependencies
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
pydantic>=1.10.7,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# HTTP and networking
requests>=2.28.2,<3.0.0
httpx>=0.24.0,<0.25.0
aiohttp>=3.8.4,<4.0.0

# XML-RPC for Odoo is part of the standard library (xmlrpc.client)

# Security
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
bcrypt>=4.0.1,<5.0.0

# System monitoring
psutil>=5.9.5,<6.0.0

# Utilities
python-multipart>=0.0.6,<0.1.0
email-validator>=2.0.0,<3.0.0
ujson>=5.7.0,<6.0.0
orjson>=3.8.12,<4.0.0

# Anthropic API for AI agent
anthropic>=0.3.0,<0.4.0

# Testing
pytest>=7.3.1,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
httpx>=0.24.0,<0.25.0

# Documentation
mkdocs>=1.4.3,<2.0.0
mkdocs-material>=9.1.6,<10.0.0
